requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.6.3
distro==1.9.0
duckduckgo_search==5.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.23.2
orjson==3.10.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyzmq==26.0.2
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
